requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
colorlog==4.2.1
concurrent-log-handler==0.9.9
elasticsearch==7.8.1
har2case==0.3.1
HttpRunner==1.5.8
idna==2.10
importlib-metadata==1.7.0
Jinja2==2.11.2
jsonpath==0.82
kafka-python==1.4.6
MarkupSafe==1.1.1
more-itertools==8.4.0
nb-log==3.4
pika==1.1.0
pluggy==0.13.1
py==1.9.0
pygame==1.9.6
pymongo==3.5.1
PyMySQL==0.9.3
pytest==4.5.0
pywin32==223
PyYAML==5.3.1
requests==2.24.0
requests-toolbelt==0.9.1
selenium==2.53.2
six==1.15.0
tomorrow3==1.1.0
urllib3==1.25.10
wcwidth==0.2.5
zipp==3.1.0
